% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographies.R
\name{geographies}
\alias{geographies}
\title{Geographies list}
\usage{
geographies(simplify = TRUE, user_key = get_user_key())
}
\arguments{
\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
List of all available geographies
}
