% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{meter_readings}
\alias{meter_readings}
\title{Meter readings list}
\usage{
meter_readings(
  geography_id,
  well_id,
  simplify = TRUE,
  user_key = get_user_key()
)
}
\arguments{
\item{geography_id}{Geography ID}

\item{well_id}{Well ID}

\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
Search all meter readings for a specified geography and well
}
