% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\name{parcels_features}
\alias{parcels_features}
\title{Parcels feature collection}
\usage{
parcels_features(
  geography_id,
  sf = TRUE,
  simplify = TRUE,
  user_key = get_user_key()
)
}
\arguments{
\item{geography_id}{Geography ID}

\item{sf}{Return object of class sf (requires sf package)}

\item{simplify}{Coerce to vector, data frame, or matrix. Ignored when sf = TRUE.}

\item{user_key}{Qanat user key}
}
\description{
List parcels as a feature collection for a specified geography
}
