% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\name{parcels_search}
\alias{parcels_search}
\title{Parcels search}
\usage{
parcels_search(
  geography_id,
  parcel_number,
  simplify = TRUE,
  user_key = get_user_key()
)
}
\arguments{
\item{geography_id}{Geography ID}

\item{parcel_number}{Parcel number (minimum first 3 characters to search)}

\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
Search all available parcel numbers for a specified geography and (partial) parcel number
}
