% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-periods.R
\name{reporting_periods}
\alias{reporting_periods}
\title{Reporting periods list}
\usage{
reporting_periods(geography_id, simplify = TRUE, user_key = get_user_key())
}
\arguments{
\item{geography_id}{Geography ID}

\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
List all reporting periods for a specified geography
}
