% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_user_key}
\alias{set_user_key}
\alias{get_user_key}
\title{Set or get Qanat API key}
\usage{
set_user_key(user_key = NULL)

get_user_key()
}
\arguments{
\item{user_key}{Qanat user key}
}
\description{
Set or get Qanat API key as environmental variable
}
