% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage-locations.R
\name{usage_locations}
\alias{usage_locations}
\title{Usage locations list}
\usage{
usage_locations(geography_id, year, simplify = TRUE, user_key = get_user_key())
}
\arguments{
\item{geography_id}{Geography ID}

\item{year}{Four-digit year}

\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
List all usage locations for a specified geography and reporting period
}
