% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage-locations.R
\name{usage_locations_features}
\alias{usage_locations_features}
\title{Usage locations feature collection}
\usage{
usage_locations_features(
  geography_id,
  year,
  sf = TRUE,
  simplify = TRUE,
  user_key = get_user_key()
)
}
\arguments{
\item{geography_id}{Geography ID}

\item{year}{Four-digit year}

\item{sf}{Return object of class sf (requires sf package)}

\item{simplify}{Coerce to vector, data frame, or matrix. Ignored when sf = TRUE.}

\item{user_key}{Qanat user key}
}
\description{
List all usage locations as a feature collection for a specified geography and reporting period
}
