% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water-accounts.R
\name{water_accounts}
\alias{water_accounts}
\title{Water accounts list}
\usage{
water_accounts(geography_id, simplify = TRUE, user_key = get_user_key())
}
\arguments{
\item{geography_id}{Geography ID}

\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
List all water accounts for a specified geography
}
