% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water-measurements.R
\name{water_measurement_types}
\alias{water_measurement_types}
\title{Water measurement types list}
\usage{
water_measurement_types(
  geography_id,
  simplify = TRUE,
  user_key = get_user_key()
)
}
\arguments{
\item{geography_id}{Geography ID}

\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
List all water measurement types for a specified geography
}
