% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water-measurements.R
\name{water_measurements}
\alias{water_measurements}
\title{Water measurements list}
\usage{
water_measurements(
  geography_id,
  parcel_id = NULL,
  water_measurement_type_id = NULL,
  year = NULL,
  simplify = TRUE,
  user_key = get_user_key()
)
}
\arguments{
\item{geography_id}{Geography ID}

\item{parcel_id}{Parcel ID}

\item{water_measurement_type_id}{Water measurement type ID}

\item{year}{Four-digit year}

\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
List all water measurements for a specified geography and parcel or
specified geography, water measurement type, and year.
If parcel_id is not specified, water measurement type and year must be specified.
}
