% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{wells}
\alias{wells}
\title{Wells list}
\usage{
wells(geography_id, simplify = TRUE, user_key = get_user_key())
}
\arguments{
\item{geography_id}{Geography ID}

\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
List all wells for a specified geography
}
