% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\name{parcels}
\alias{parcels}
\title{Parcels list}
\usage{
parcels(geography_id, simplify = TRUE, user_key = get_user_key())
}
\arguments{
\item{geography_id}{Geography ID}

\item{simplify}{Coerce to vector, data frame, or matrix}

\item{user_key}{Qanat user key}
}
\description{
List all parcel numbers for a specified geography
}
